readded Val_some
