+++ src/Native/dialog.c
@@ -7,15 +7,6 @@
 
 #include "caml_values.h"
 
-static inline value Val_some(long v) {
-  (void)Val_some;
-  CAMLparam1(v);
-  CAMLlocal1(some);
-  some = caml_alloc(1, 0);
-  Store_field(some, 0, v);
-  CAMLreturn(some);
-}
-
 #include <string.h>
 
 #ifdef WIN32
