+++ src/Native/utilities.c
@@ -1,8 +1,10 @@
-#include "caml_values.h"
 #include <caml/callback.h>
 #include <caml/memory.h>
 #include <caml/mlvalues.h>
 #include <caml/threads.h>
+#include <caml/alloc.h>
+
+#include "caml_values.h"
 
 /* Sourced from Brisk's `BriskCocoa`
     Thank you @wokalski!
